Support both Pydantic v1 and v2
This is another attempt at supporting both Pydantic v1 and v2, but
without using the v1 module. Instead we configure WebAuthnBaseModel
slightly differently based on the pydantic version.

This contains a slight behavior change in that trailing = in base64
encoded values are not removed, because pydantic v2 doesn't support
that. If that is a problem (I'm not familiar with the spec) I can try to
see if I can remove them somehow by doing some post-processing of the
serialized values provided by pydantic