authentication: return user verified flag
Some other libs like webauthn-rs enforce that when user verification is
only preferred and that a device previously performed user verification,
it must continue to perform it on later authentication, as it proved
that it was user verification capable. Whether this out-of-spec check is
desirable is an open question, however if someone wants to align
behavior between a webauthn-rs-based implem and a py_webauthn implem
then returning this information is necessary.