Replace setup.py with declarative pyproject.toml
While setup.py still works, it is somewhat old-school for pure python
projects. Replace it with a pyproject.toml which is a nice declarative
file. See the Python Packaging User Guide for more details:
https://packaging.python.org/en/latest/guides/modernize-setup-py-project/

The `license` metadata was updated from the ambiguous "BSD" which is
rejected by recent setuptools, to the SPDX license identifier
"BSD-3-Clause" (matching the LICENSE file). The license classifier
"License :: OSI Approved :: BSD License" was removed because setuptools
doesn't allow to have both.

The "download URL" metadata is removed, because the `{version}`
templating would no longer work, and it doesn't seem very important.

I checked the wheel and it looks good.

The publish workflow is updated to avoid the use of setup.py, using the
`build` package instead. See:
https://packaging.python.org/en/latest/tutorials/packaging-projects/#generating-distribution-archives